Fixes session lifecycle callbacks to run in call-loop and correctly grab session (#51)
* Fixes session lifecycle callbacks to run in call-loop and correctly grab session

Previously, the session lifecycle callbacks were not async,
resulting in errors inside of them to break login/logout if the
application had invalid logic inside of the callbacks. This commit
places the callbacks inside of the run-loop so that they become
async. It also correctly grabs the session inside of the login
route before executing the callbacks.

* bump version number